+++ src/components/features/composite/color/color.tsx
@@ -10,12 +10,15 @@ const Light: FunctionComponent<ColorProps> = (props) => {
     const { deviceState, feature, onChange, minimal } = props;
     const value = {};
     for (const innerFeature of feature.features) {
-        value[innerFeature.name] =
-            (deviceState[feature.property] as Record<string, Record<string, unknown>>)?.[innerFeature.property] ?? 0;
+        // just in case the number comes in as string
+        const propValue = Number.parseFloat(
+            (deviceState[feature.property] as Record<string, Record<string, unknown>>)?.[innerFeature.name],
+        );
+        value[innerFeature.name] = Number.isNaN(propValue) ? 0 : propValue;
     }
     return (
         <ColorEditor
-            onChange={(color) => onChange(feature.endpoint as Endpoint, { color })}
+            onChange={(color) => onChange(feature.endpoint as Endpoint, { [feature.property ?? 'color']: color })}
             value={value as AnyColor}
             format={feature.name}
             minimal={minimal}